refactor(web): split DirectTerminal.tsx into focused modules
Closes #769.

DirectTerminal.tsx was 951 lines — 2.4x over the 400-line component limit
(C-04) — with theme data, font helpers, the xterm setup effect, the
fullscreen resize effect, clipboard handlers, and the entire chrome UI
bundled into one file. The April 2026 audit comment on the issue
confirmed the UI revamp had landed and the structural split was still
wanted.

Extracted into components/terminal/:
- terminal-themes.ts — dark/light ITheme builder
- terminal-font.ts — font size constants, getStoredFontSize,
  resolveMonoFontFamily
- terminal-clipboard.ts — XDA (CSI > q) handler, OSC 52 decoder, copy
  keybinding
- useXtermTerminal.ts — owns the xterm instance, mux wiring, selection
  preservation, touch scroll, and the reconnect/theme/font-size effects
- useFullscreenResize.ts — the RAF-polling fullscreen resize effect
- TerminalControls.tsx — chromed bar + chromeless floating controls
  with OpenCode reload logic

DirectTerminal.tsx is now 130 lines of composition. buildTerminalThemes
and resolveMonoFontFamily are re-exported so existing tests continue to
import them from @/components/DirectTerminal.

The ~280-line xterm setup effect in useXtermTerminal stays intact; its
selection buffering, programmatic-scroll tracking, and cleanup ordering
are too tightly coupled to split further without passing refs between
6 hooks. Further splitting would add indirection without reducing
complexity.
